requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
pandas==2.1.1
openai==1.14.3
